+++ src/providers/identity/index.ts
@@ -1,0 +1,8 @@
+export { Auth0Adapter } from "./auth0/adapter.js";
+export { DecentroAdapter } from "./decentro/adapter.js";
+export { DigioAdapter } from "./digio/adapter.js";
+export { HyperVergeAdapter } from "./hyperverge/adapter.js";
+export { IdfyAdapter } from "./idfy/adapter.js";
+export { KarzaAdapter } from "./karza/adapter.js";
+export { PerfiosAdapter } from "./perfios/adapter.js";
+export { SetuAdapter } from "./setu/adapter.js";